# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RelancePro Africa API",
    "description": "API publique pour la gestion des créances clients en Afrique. Support REST et GraphQL.",
    "version": "1.0.0",
    "contact": {
      "name": "RelancePro Africa Support",
      "email": "support@relancepro.africa"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.relancepro.africa",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Utilisez votre clé API comme Bearer token. Exemple: Authorization: Bearer rpa_live_xxxxx"
      }
    },
    "schemas": {
      "Client": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "company": { "type": "string", "nullable": true },
          "address": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "riskScore": { "type": "integer", "nullable": true },
          "riskLevel": { "type": "string", "enum": ["low", "medium", "high"], "nullable": true },
          "status": { "type": "string", "enum": ["active", "inactive", "blacklisted"] },
          "debtsCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Debt": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "reference": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "amount": { "type": "number" },
          "currency": { "type": "string" },
          "paidAmount": { "type": "number" },
          "remainingAmount": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "paid", "partial", "disputed", "cancelled"] },
          "issueDate": { "type": "string", "format": "date-time", "nullable": true },
          "dueDate": { "type": "string", "format": "date-time" },
          "paidDate": { "type": "string", "format": "date-time", "nullable": true },
          "reminderCount": { "type": "integer" },
          "lastReminderAt": { "type": "string", "format": "date-time", "nullable": true },
          "paymentProbability": { "type": "number", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object", "nullable": true }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/clients/{id}": {
      "get": {
        "tags": ["Clients"],
        "summary": "Get client details",
        "operationId": "getClient",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Client" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "put": {
        "tags": ["Clients"],
        "summary": "Update client",
        "operationId": "updateClient",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "company": { "type": "string" },
                  "address": { "type": "string" },
                  "notes": { "type": "string" },
                  "status": { "type": "string", "enum": ["active", "inactive", "blacklisted"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Client" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "delete": {
        "tags": ["Clients"],
        "summary": "Delete client",
        "operationId": "deleteClient",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Client deleted" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "409": { "description": "Cannot delete client with existing debts" }
        }
      }
    },
...
    "/api/v1/debts/{id}": {
      "get": {
        "tags": ["Debts"],
        "summary": "Get debt details",
        "operationId": "getDebt",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Debt" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "put": {
        "tags": ["Debts"],
        "summary": "Update debt",
        "operationId": "updateDebt",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reference": { "type": "string" },
                  "description": { "type": "string" },
                  "amount": { "type": "number" },
                  "currency": { "type": "string" },
                  "dueDate": { "type": "string", "format": "date-time" },
                  "status": { "type": "string", "enum": ["pending", "paid", "partial", "disputed", "cancelled"] },
                  "paidAmount": { "type": "number" },
                  "paidDate": { "type": "string", "format": "date-time" },
                  "autoRemindEnabled": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Debt updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Debt" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "delete": {
        "tags": ["Debts"],
        "summary": "Delete debt",
        "operationId": "deleteDebt",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Debt deleted" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
...
  },
...
}